before get vaccination

+++ src/component/header/Header.jsx
@@ -4,7 +4,7 @@ import logo from '../../assets/logo/logo.png';
 import 'bootstrap/dist/css/bootstrap.min.css';
 // import '../../assets/fonts/Font.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faUserAlt } from '@fortawesome/free-solid-svg-icons';
+import { faChevronDown, faUserAlt } from '@fortawesome/free-solid-svg-icons';
 
 class Header extends React.Component{
     constructor(props){
@@ -20,20 +20,20 @@ class Header extends React.Component{
         }
     }
     componentDidMount(){
-        var isConnected = false;
+        // var type = 0;
+        // if(localStorage.getItem('typeConnected')!==null&&localStorage.getItem('typeConnected')!==undefined&&localStorage.getItem('typeConnected')!=="")
+        //     type = localStorage.getItem('typeConnected');
         const chemin = window.location.pathname.split('/')[1];
         this.setState({way:chemin});
         // alert(chemin)
         // const isConnected = localStorage.getItem('connected');
-        if(localStorage.getItem('idCentre')===null && localStorage.getItem('idUser')===null){
-            isConnected = false;
-            if(chemin==='profil'||chemin==='profil-centre'){
+        if(localStorage.getItem('idCentre')===null && localStorage.getItem('idUser')===null && localStorage.getItem('idStaff')===null){
+            if(chemin==='profil'||chemin==='profil-centre'||chemin==='profil-staff'){
                 alert('Vous n\'êtes pas autororisé à accéder ce lien');
                 window.location.replace('/');
             }
-        }else if(localStorage.getItem('idCentre')!==null && localStorage.getItem('idUser')===null){
-            isConnected = true;
-            if(chemin==='profil'){
+        }else if(localStorage.getItem('idCentre')!==null && localStorage.getItem('idUser')===null && localStorage.getItem('idStaff')===null){
+            if(chemin==='profil'||chemin==='profil-staff'){
                 alert('Vous n\'êtes pas autororisé à accéder ce lien');
                 window.location.replace('/profil-centre/1');
             }
@@ -42,9 +42,8 @@ class Header extends React.Component{
                 window.location.replace('/profil-centre/1');
             }
             this.setState({type:'centre', pseudo:localStorage.getItem('pseudo'), photo: localStorage.getItem('photo')});
-        }else if(localStorage.getItem('idCentre')===null && localStorage.getItem('idUser')!==null){
-            isConnected = true;
-            if(chemin==='profil-centre'){
+        }else if(localStorage.getItem('idCentre')===null && localStorage.getItem('idUser')!==null && localStorage.getItem('idStaff')===null){
+            if(chemin==='profil-centre'||chemin==='profil-staff'){
                 alert('Vous n\'êtes pas autororisé à accéder ce lien');
                 window.location.replace('/');
             }
@@ -53,6 +52,16 @@ class Header extends React.Component{
                 window.location.replace('/profil');
             }
             this.setState({type:'patient', pseudo:localStorage.getItem('pseudo'), photo: localStorage.getItem('photo')});
+        }else if(localStorage.getItem('idCentre')===null && localStorage.getItem('idUser')===null && localStorage.getItem('idStaff')!==null){
+            if(chemin==='profil'||chemin==='profil-centre'){
+                alert('Vous n\'êtes pas autororisé à accéder ce lien');
+                window.location.replace('/profil-staff/1');
+            }
+            if(chemin==='connexion-centre' || chemin==='inscription-centre' || chemin==='connexion' || chemin==='inscription'){
+                alert('Vous êtes déjà connecté');
+                window.location.replace('/profil-staff/1');
+            }
+            this.setState({type:'staff', pseudo:localStorage.getItem('pseudo'), photo: 'profile.png'});
         }
         // console.log(localStorage.getItem('connected')===null);
     }
@@ -78,6 +87,21 @@ class Header extends React.Component{
                 return(
                     <>
                         <li onClick={()=>{window.location.replace("/profil")}}>Compte</li>
+                        <li onClick={()=>{localStorage.clear(); window.location.replace("/connexion")}}>Déconnexion</li>
+                    </>
+                )
+            }else{
+                return(
+                    <>
+                        <li onClick={()=>{localStorage.clear(); window.location.replace("/connexion")}}>Déconnexion</li>
+                    </>
+                )
+            }
+        }else if(this.state.type==='staff'){
+            if(this.state.way!=='profil-staff'){
+                return(
+                    <>
+                        <li onClick={()=>{window.location.replace("/profil-staff/1")}}>Compte</li>
                         <li onClick={()=>{localStorage.clear(); window.location.replace("/connexion")}}>Déconnexion</li>
                     </>
                 )
